change routes to use /api/v1
