+++ config/routes.rb
@@ -1,7 +1,8 @@
 Rails.application.routes.draw do
-  resources :posts
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
-
-  # Defines the root path route ("/")
-  # root "articles#index"
+  namespace :api do
+    namespace :v1 do
+      resources :posts
+    end
+  end
 end
